+++ src/helpers/functions/Role.js
@@ -4,6 +4,8 @@ exports.getRoles = () => JSON.parse(JSON.stringify(roles));
 
 exports.getVillagerRoles = () => JSON.parse(JSON.stringify(roles.filter(({ side }) => side === "villagers")));
 
+exports.getWerewolfRoles = () => JSON.parse(JSON.stringify(roles.filter(({ side }) => side === "werewolves")));
+
 exports.getRoleNames = () => JSON.parse(JSON.stringify(roleNames));
 
 exports.getSideNames = () => JSON.parse(JSON.stringify(sideNames));