+++ tests/e2e/game/index.js
@@ -21,4 +21,5 @@ describe("E2E - 🎲 Game tests", () => {
     require("./game-won-by-white-werewolf.test");
     require("./endless-votes.test");
     require("./game-won-by-angel.test");
+    require("./thief-steals.test");
 });