+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 * [#119](https://github.com/antoinezanardi/werewolves-assistant-api/issues/119) - Implement White Werewolf role.
 * [#120](https://github.com/antoinezanardi/werewolves-assistant-api/issues/120) - Implement Stuttering Judge role.
 * [#126](https://github.com/antoinezanardi/werewolves-assistant-api/issues/126) - Implement Angel role.
+* [#127](https://github.com/antoinezanardi/werewolves-assistant-api/issues/127) - Implement Thief Role.
 
 ### 🌟 Enhancements
 
@@ -41,6 +42,7 @@
 * `@sentry/node` updated to version `6.1.0`.
 * `chai` updated to version `4.3.0`.
 * `express-rate-limit` updated to version `5.2.5`.
+* `mocha` updated to version `8.3.0`.
 * `mongoose` updated to version `5.11.15`.
 
 ---